--- cookies.txt
@@ -1,10 +1,0 @@
-# Netscape HTTP Cookie File
-# This file is generated by yt-dlp.  Do not edit.
-
-.youtube.com	TRUE	/	TRUE	1763885783	GPS	1
-.youtube.com	TRUE	/	FALSE	0	PREF	hl=en&tz=UTC
-.youtube.com	TRUE	/	TRUE	0	SOCS	CAI
-.youtube.com	TRUE	/	TRUE	1779435983	VISITOR_INFO1_LIVE	bSnGf1bl7rQ
-.youtube.com	TRUE	/	TRUE	1779435983	VISITOR_PRIVACY_METADATA	CgJJThIEGgAgFg%3D%3D
-.youtube.com	TRUE	/	TRUE	0	YSC	IPLyevpgbEI
-.youtube.com	TRUE	/	TRUE	1779435983	__Secure-ROLLOUT_TOKEN	CPGYpefj-snowwEQjYiH8uOHkQMYjYiH8uOHkQM%3D
